Add files via upload
Assignment 1 for CSCI 121 Data Structures. Create a money converter. Creativity portion: User can input a $ or not. Give user the option to run a trial transaction. Allows them to input the amount of money they have, and the amount of a product, and giving the coinage difference. As well as the option to input a second amount.